fix: recreate missing components and fix TypeScript errors - Recreate Switch.tsx, Tabs.tsx, ClinicalAgent.ts, MCPContextViewer.tsx - Fix 45+ TypeScript errors - Fix linting issues - All tests passing (229/229) - Ready for CI/CD pipeline

+++ src/services/CryptoService.ts
@@ -77,7 +77,7 @@ export class CryptoService {
   }
 
   private static generateIV(): Uint8Array {
-    return window.crypto.getRandomValues(new Uint8Array(this.IV_LENGTH));
+    return window.crypto.getRandomValues(new Uint8Array(this.IV_LENGTH) as Uint8Array);
   }
 
   private static async encryptData(data: string, key: CryptoKey): Promise<EncryptedData> {
